refactor: unifying client mocks
